Remove print/eprint snippet.
This removes the print! and eprint! snippets. They are used much less often
(about 5% of the time on crates.io), and often get in the way of wanting to
complete the `ln` versions. Snippets aren't really a replacement for full
completion, and I think should be reserved for common templates. If the user
really wants the non-`ln` version, they can just type it manually, or delete
the `ln` after completion. If the user really wants these snippets, they
can be added back into their User directory.