Next/prev jump to end of span.
This changes the next/prev commands to jump to the end of a message span instead
of the beginning. When using inline phantoms, jumping to the beginning can
be hard to see the message for a multi-line message.

I've flip-flopped over time which I like better. I've been using "end" for a
while now, and it seems to work better for me. If anyone objects, we can always
add a config setting.

This also changes how the test works, since the old test didn't handle this new
format very well.